Fix ChangeActivityState from activity with timerEvents to any other Separate related changeActivityState tests from RuntimeServiceTests into RuntimeServiceChangeStateTest
