+++ modules/flowable-engine/src/test/resources/org/flowable/engine/test/api/oneTaskSubProcessWithTimer.bpmn20.xml
@@ -1,47 +1,46 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<definitions id="definitions" xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
-  xmlns:activiti="http://activiti.org/bpmn" targetNamespace="Examples">
+<definitions id="definitions" xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" targetNamespace="Examples">
 
-  <process id="startSimpleSubProcess">
+    <process id="startSimpleSubProcess">
 
-    <startEvent id="theStart" />
+        <startEvent id="theStart"/>
 
-    <sequenceFlow id="flow1" sourceRef="theStart" targetRef="taskBefore" />
-    
-    <userTask id="taskBefore" name="Task before sub process" />
-    
-    <sequenceFlow id="flow2" sourceRef="taskBefore" targetRef="subProcess" />
+        <sequenceFlow id="flow1" sourceRef="theStart" targetRef="taskBefore"/>
 
-    <subProcess id="subProcess">
+        <userTask id="taskBefore" name="Task before sub process"/>
 
-      <startEvent id="theSubProcessStart" />
+        <sequenceFlow id="flow2" sourceRef="taskBefore" targetRef="subProcess"/>
 
-      <sequenceFlow id="subflow1" sourceRef="theSubProcessStart" targetRef="subtask" />
+        <subProcess id="subProcess">
 
-      <userTask id="subtask" name="Task in subprocess" />
+            <startEvent id="theSubProcessStart"/>
 
-      <sequenceFlow id="subflow2" sourceRef="subtask" targetRef="theSubProcessEnd" />
+            <sequenceFlow id="subflow1" sourceRef="theSubProcessStart" targetRef="subTask"/>
 
-      <endEvent id="theSubProcessEnd" />
+            <userTask id="subTask" name="Task in subprocess"/>
 
-    </subProcess>
-    
-    <boundaryEvent id="boundaryTimerEvent" attachedToRef="subProcess">
-      <timerEventDefinition>
-        <timeDuration>PT1H</timeDuration>
-      </timerEventDefinition>
-    </boundaryEvent>
-    
-    <sequenceFlow id="flow3" sourceRef="boundaryTimerEvent" targetRef="theEnd" />
+            <sequenceFlow id="subflow2" sourceRef="subTask" targetRef="theSubProcessEnd"/>
 
-    <sequenceFlow id="flow4" sourceRef="subProcess" targetRef="taskAfter" />
-    
-    <userTask id="taskAfter" name="Task after sub process" />
-    
-    <sequenceFlow id="flow5" sourceRef="taskAfter" targetRef="theEnd" />
+            <endEvent id="theSubProcessEnd"/>
 
-    <endEvent id="theEnd" />
+        </subProcess>
 
-  </process>
+        <boundaryEvent id="boundaryTimerEvent" attachedToRef="subProcess">
+            <timerEventDefinition>
+                <timeDuration>PT1H</timeDuration>
+            </timerEventDefinition>
+        </boundaryEvent>
+
+        <sequenceFlow id="flow3" sourceRef="boundaryTimerEvent" targetRef="theEnd"/>
+
+        <sequenceFlow id="flow4" sourceRef="subProcess" targetRef="taskAfter"/>
+
+        <userTask id="taskAfter" name="Task after sub process"/>
+
+        <sequenceFlow id="flow5" sourceRef="taskAfter" targetRef="theEnd"/>
+
+        <endEvent id="theEnd"/>
+
+    </process>
 
 </definitions>
